requirements.txt
gradio
pyyaml
plotly
trimesh
scipy